添加离线消息同步
引入了客户端离线消息同步管理器和基于本地文件的消息存储，包括批量消息拉取和同步序列管理。增强了HTTP、TCP和WebSocket客户端，以支持离线同步API、批量ACK和本地同步点更新。添加了Redis配置加载器和相关属性，以备将来与Redis集成。更新了.gitignore和文档，以适应新功能和本地数据处理。

+++ im-message-server/src/main/java/com/vanky/im/message/constant/MessageConstants.java
@@ -123,19 +123,19 @@ public class MessageConstants {
     // ========== 消息推送状态常量 ==========
 
     /** 消息状态：已发送（等待推送） */
-    public static final int MESSAGE_STATUS_SENT = 0;
+    public static final byte MESSAGE_STATUS_SENT = 0;
 
     /** 消息状态：推送成功（客户端已确认） */
-    public static final int MESSAGE_STATUS_DELIVERED = 1;
+    public static final byte MESSAGE_STATUS_DELIVERED = 1;
 
     /** 消息状态：已读 */
-    public static final int MESSAGE_STATUS_READ = 2;
+    public static final byte MESSAGE_STATUS_READ = 2;
 
     /** 消息状态：撤回 */
-    public static final int MESSAGE_STATUS_RECALLED = 3;
+    public static final byte MESSAGE_STATUS_RECALLED = 3;
 
     /** 消息状态：推送失败 */
-    public static final int MESSAGE_STATUS_FAILED = 4;
+    public static final byte MESSAGE_STATUS_FAILED = 4;
 
     // ========== 私有构造函数 ==========
